TEST: CICD

+++ test/cleanup.sh
@@ -3,8 +3,22 @@
 #   Test cleanup script
 #
 
-killall -q fastProgram
-killall -q proxyServer
+echo IN CLEANUP
+env | grep TESTME
+
+if [ "${TESTME_SUCCESS}" = "1" ]; then
+    rm -f appweb.log
+else
+    echo "Appweb log:"
+    echo "--------------------------------"
+    cat appweb.log
+    echo "--------------------------------"
+fi
+
+if [ "$TESTME_OS" != "windows" ] ; then
+    killall -q fastProgram
+    killall -q proxyServer
+fi
 
 rm -f *.log
 rm -f .test-prepared
